Fixed warnings

+++ src_cpp/EXT/EXT_suite.cpp
@@ -49,7 +49,7 @@ goods and services.
 */
 
 #if defined MPI1 || defined NBC || defined MPIIO || defined RMA
-#error Legacy benchmark components can't be linked together
+#error Legacy benchmark components cant be linked together
 #endif
 
 #include <set>
@@ -474,7 +474,7 @@ template <> bool BenchmarkSuite<BS_EXT>::prepare(const args_parser &parser, cons
     IMB_internal_barrier = (parser.get<bool>("imb_barrier") ? 1 : 0);
 
     int alignment = parser.get<int>("alignment");
-    if (alignment < sizeof(void*)) {
+    if (alignment < (int)sizeof(void*)) {
         alignment = sizeof(void*);
     }
     int power2 = 1;
